move PrefixedWithPaddingBefore to construct_extensions
